📝 [api] fix scramble docs drifting from actual response shapes
ShowUserAction's stale @response User docblock resolved via the
file's use imports to the raw App\Models\User instead of UserResource
(the .ai/rules/controllers.md gotcha), so the docs claimed GET /user
returns a bare User with no envelope when the real response is
{data: User}. Removing it lets Scramble infer from the
UserResource::make(...) return statement, which also collapses a
colliding duplicate User/App.Http.Resources.User schema pair back
into one correctly-referenced User schema.

Act's appreciates_count/comments_count were documented as
anyOf[string, integer(enum 0)] since Scramble can't type the dynamic
withCount() attribute through `?? 0`. Casting to (int) makes the docs
match reality: these are always plain integers.

+++ app/Http/Resources/Act.php
@@ -30,8 +30,8 @@ class Act extends JsonResource
             'type' => $this->type,
             'created_at' => $this->created_at,
             'updated_at' => $this->updated_at,
-            'appreciates_count' => $this->appreciates_count ?? 0,
-            'comments_count' => $this->comments_count ?? 0,
+            'appreciates_count' => (int) ($this->appreciates_count ?? 0),
+            'comments_count' => (int) ($this->comments_count ?? 0),
             'appreciates' => AppreciateResource::collection($this->whenLoaded('appreciates')),
             'comments' => CommentResource::collection($this->whenLoaded('comments')),
         ];
